Version: 2001-01-01a 93k+
Sourced from ast-open-history version 2000-10-31 and:
07466222abec22ca7ccded93737d750c4632138a0febb3f130a9ce1835e4ac53  INIT.2001-01-01a.tgz
80f41374257084d67e89e04bc34fb1eb3c2e3d37af81223c5f35fdd35b661b51  ast-base.2001-01-01a.tgz

+++ src/cmd/tests/sfio/tsetbuf.c
@@ -1,7 +1,7 @@
 /*******************************************************************
 *                                                                  *
 *             This software is part of the ast package             *
-*                Copyright (c) 1999-2000 AT&T Corp.                *
+*                Copyright (c) 1999-2001 AT&T Corp.                *
 *        and it may only be used by you under license from         *
 *                       AT&T Corp. ("AT&T")                        *
 *         A copy of the Source Code Agreement is available         *
@@ -20,7 +20,6 @@
 *                         Florham Park NJ                          *
 *                                                                  *
 *               Glenn Fowler <gsf@research.att.com>                *
-*                                                                  *
 *******************************************************************/
 #include	"sftest.h"
 #ifdef SF_APPEND
@@ -61,7 +60,7 @@ MAIN()
 
 	if(write(fd,buf,sizeof(buf)) != sizeof(buf))
 		terror("Writing to file\n");
-	if(lseek(fd,0L,0) < 0)
+	if(lseek(fd, (off_t)0, 0) < 0)
 		terror("Seeking back to origin\n");
 
 	if(!(f = sfnew((Sfio_t*)0,buf,sizeof(buf),fd,SF_WRITE)))
